fix axiosInstance and not appear footer in learn lesson page

+++ src/services/axiosInstance.ts
@@ -4,6 +4,8 @@ import { paths, roles } from "../consts";
 import { message } from "antd";
 import { getUserFromLocalStorage } from "./auth.ts";
 
+let isTokenExpired = false;
+
 export const axiosInstance = axios.create({
   baseURL: config.BASE_URL,
   headers: {
@@ -15,12 +17,10 @@ export const axiosInstance = axios.create({
 
 axiosInstance.interceptors.request.use(
   (config) => {
-    if (!config.headers.skipAuth) {
       const token = localStorage.getItem("token");
       if (token) {
         config.headers.Authorization = `Bearer ${token}`;
       }
-    }
     return config;
   },
   (error) => {
@@ -42,35 +42,40 @@ axiosInstance.interceptors.response.use(
     if (error.response) {
       const { data } = error.response;
       console.log(error.response);
-      if(data.message && !data.errors)
-      switch (error.response.status) {
-        case 403: {
-          message.error(data.message);
-          const user = getUserFromLocalStorage();
-          setTimeout(() => {
-            if (user.role === roles.ADMIN) {
-              window.location.href = paths.ADMIN_LOGIN;
-            } else {
-              window.location.href = paths.LOGIN;
+      if(data.message.length > 0) {
+        switch (error.response.status) {
+          case 403: {
+            if(!isTokenExpired){
+              isTokenExpired = true
+              message.error(data.message);
+              const user = getUserFromLocalStorage();
+              setTimeout(() => {
+                if (user.role === roles.ADMIN) {
+                  window.location.href = paths.ADMIN_LOGIN;
+                } else {
+                  window.location.href = paths.LOGIN;
+                }
+                localStorage.clear();
+                isTokenExpired = false;
+              }, 1300);
             }
-            localStorage.clear();
-          }, 1300);
-          break;
+            break;
+          }
+
+          case 404:
+            message.error(data.message);
+            window.location.href = paths.NOTFOUND;
+            break;
+
+          case 500:
+            message.error(data.message);
+            window.location.href = paths.INTERNAL_SERVER_ERROR;
+            break;
+
+          default:
+            message.error(data.message);
+            break;
         }
-
-        case 404:
-          message.error(data.message);
-          window.location.href = paths.NOTFOUND;
-          break;
-
-        case 500:
-          message.error(data.message);
-          window.location.href = paths.INTERNAL_SERVER_ERROR;
-          break;
-
-        default:
-          message.error(data.message);
-          break;
       }
 
       else{
